+++ repo/src/main/java/com/parashift/onlyoffice/Prepare.java
@@ -1,13 +1,17 @@
 package com.parashift.onlyoffice;
 
 import org.alfresco.model.ContentModel;
-import org.alfresco.repo.admin.SysAdminParams;
-import org.alfresco.service.cmr.repository.ContentData;
-import org.alfresco.service.cmr.repository.NodeRef;
-import org.alfresco.service.cmr.repository.NodeService;
-import org.alfresco.service.cmr.security.AuthenticationService;
+import org.alfresco.repo.content.MimetypeMap;
+import org.alfresco.service.cmr.repository.*;
 import org.alfresco.service.namespace.QName;
-import org.alfresco.util.UrlUtil;
+import org.apache.poi.hslf.HSLFSlideShow;
+import org.apache.poi.hssf.usermodel.HSSFSheet;
+import org.apache.poi.hssf.usermodel.HSSFWorkbook;
+import org.apache.poi.hwpf.HWPFDocument;
+import org.apache.poi.xslf.usermodel.XMLSlideShow;
+import org.apache.poi.xssf.usermodel.XSSFSheet;
+import org.apache.poi.xssf.usermodel.XSSFWorkbook;
+import org.apache.poi.xwpf.usermodel.XWPFDocument;
 import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,10 +22,7 @@ import org.springframework.extensions.webscripts.WebScriptResponse;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.Resource;
-import java.io.IOException;
-import java.io.Serializable;
-import java.io.Writer;
-import java.text.SimpleDateFormat;
+import java.io.*;
 import java.util.Map;
 import java.util.Properties;
 
@@ -39,6 +40,9 @@ public class Prepare extends AbstractWebScript {
 
     @Autowired
     NodeService nodeService;
+
+    @Autowired
+    ContentService contentService;
 
     @Resource(name = "global-properties")
     Properties globalProp;
@@ -71,6 +75,7 @@ public class Prepare extends AbstractWebScript {
                     responseJson.put("key", key);
                     responseJson.put("docTitle", properties.get(ContentModel.PROP_NAME));
                     responseJson.put("mimeType", contentData.getMimetype());
+                    responseJson.put("abovePreviewThreshold", checkAbovePreviewThreshold(nodeRef, contentData));
 
                     if(globalProp.containsKey("onlyoffice.lang")) {
                         responseJson.put("lang", globalProp.get("onlyoffice.lang"));
@@ -94,4 +99,125 @@ public class Prepare extends AbstractWebScript {
 
         }
     }
+
+    private boolean checkAbovePreviewThreshold(NodeRef nodeRef, ContentData contentData) {
+        boolean result = false;
+        Long documentMaxSize = Long.parseLong((String) globalProp.getOrDefault("onlyoffice.preview.document.size.threshold", "0"));
+        Long documentSize = contentData.getSize();
+
+        if (documentMaxSize != 0 && documentSize > documentMaxSize) {
+            logger.debug("Document size {} exceeds threshold {}.", documentSize, documentMaxSize);
+            result = true;
+        }
+
+        if (!result) {
+            String mimeType = contentData.getMimetype();
+
+            if (mimeType != null) {
+
+                Integer docxMaxParagraph = Integer.parseInt((String) globalProp.getOrDefault("onlyoffice.preview.docx.threshold", "0"));
+                Integer docMaxPage = Integer.parseInt((String) globalProp.getOrDefault("onlyoffice.preview.doc.threshold", "0"));
+                Integer xlsxMaxRows = Integer.parseInt((String) globalProp.getOrDefault("onlyoffice.preview.xlsx.threshold", "0"));
+                Integer xlsMaxRows = Integer.parseInt((String) globalProp.getOrDefault("onlyoffice.preview.xls.threshold", "0"));
+                Integer pptxMaxSlides = Integer.parseInt((String) globalProp.getOrDefault("onlyoffice.preview.pptx.threshold", "0"));
+                Integer pptMaxSlides = Integer.parseInt((String) globalProp.getOrDefault("onlyoffice.preview.ppt.threshold", "0"));
+
+                logger.debug("Thresholds: docx: {}, doc: {}, xlsx: {}, xls: {}, pptx: {}, ppt: {}", docxMaxParagraph, docMaxPage, xlsxMaxRows, xlsMaxRows, pptxMaxSlides, pptMaxSlides);
+
+                ContentReader contentReader = contentService.getReader(nodeRef, ContentModel.PROP_CONTENT);
+                try {
+                    switch (mimeType) {
+                        case MimetypeMap.MIMETYPE_OPENXML_WORD_TEMPLATE:
+                        case MimetypeMap.MIMETYPE_OPENXML_WORD_TEMPLATE_MACRO:
+                        case MimetypeMap.MIMETYPE_OPENXML_WORDPROCESSING_MACRO:
+                        case MimetypeMap.MIMETYPE_OPENXML_WORDPROCESSING:
+                            if (docxMaxParagraph > 0) {
+                                try (InputStream inputStream = contentReader.getContentInputStream()) {
+                                    XWPFDocument docx = new XWPFDocument(inputStream);
+                                    int paragraphNum = docx.getBodyElements().size();
+                                    logger.debug("DOCX paragraph number is: {}", paragraphNum);
+                                    result = paragraphNum > docxMaxParagraph;
+                                }
+                            }
+                            break;
+                        case MimetypeMap.MIMETYPE_WORD:
+                            if (docMaxPage > 0) {
+                                try (InputStream inputStream = contentReader.getContentInputStream()) {
+                                    HWPFDocument doc = new HWPFDocument(inputStream);
+                                    int pageCount = doc.getSummaryInformation().getPageCount();
+                                    logger.debug("DOC page count is: {}", pageCount);
+                                    result = pageCount > docMaxPage;
+                                }
+                            }
+                            break;
+                        case MimetypeMap.MIMETYPE_OPENXML_SPREADSHEET_TEMPLATE:
+                        case MimetypeMap.MIMETYPE_OPENXML_SPREADSHEET_TEMPLATE_MACRO:
+                        case MimetypeMap.MIMETYPE_OPENXML_SPREADSHEET_MACRO:
+                        case MimetypeMap.MIMETYPE_OPENXML_SPREADSHEET:
+                            if (xlsxMaxRows > 0) {
+                                try (InputStream inputStream = contentReader.getContentInputStream()) {
+                                    XSSFWorkbook xlsx = new XSSFWorkbook(inputStream);
+                                    Integer totalRows = 0;
+                                    for (int i = 0; i < xlsx.getNumberOfSheets(); i++) {
+                                        XSSFSheet sheet = xlsx.getSheetAt(i);
+                                        totalRows += sheet.getPhysicalNumberOfRows();
+                                        logger.debug("XLSX totalRows: {}", totalRows);
+                                        if (result = totalRows > xlsxMaxRows) {
+                                            break;
+                                        }
+                                    }
+                                }
+                            }
+                            break;
+                        case MimetypeMap.MIMETYPE_EXCEL:
+                            if (xlsMaxRows > 0) {
+                                try (InputStream inputStream = contentReader.getContentInputStream()) {
+                                    HSSFWorkbook xls = new HSSFWorkbook(inputStream);
+                                    Integer xlsTotalRows = 0;
+                                    for (int i = 0; i < xls.getNumberOfSheets(); i++) {
+                                        HSSFSheet sheet = xls.getSheetAt(i);
+                                        xlsTotalRows += sheet.getPhysicalNumberOfRows();
+                                        logger.debug("XLS totalRows: {}", xlsTotalRows);
+                                        if (result = xlsTotalRows > xlsMaxRows) {
+                                            break;
+                                        }
+                                    }
+                                }
+                            }
+                            break;
+                        case MimetypeMap.MIMETYPE_OPENXML_PRESENTATION_TEMPLATE:
+                        case MimetypeMap.MIMETYPE_OPENXML_PRESENTATION_TEMPLATE_MACRO:
+                        case MimetypeMap.MIMETYPE_OPENXML_PRESENTATION_SLIDESHOW:
+                        case MimetypeMap.MIMETYPE_OPENXML_PRESENTATION_SLIDESHOW_MACRO:
+                        case MimetypeMap.MIMETYPE_OPENXML_PRESENTATION_MACRO:
+                        case MimetypeMap.MIMETYPE_OPENXML_PRESENTATION:
+                            if (pptxMaxSlides > 0) {
+                                try (InputStream inputStream = contentReader.getContentInputStream()) {
+                                    XMLSlideShow pptx = new XMLSlideShow(inputStream);
+                                    int slidesNum = pptx.getSlides().length;
+                                    logger.debug("PPTX slides number is: {}", slidesNum);
+                                    result = slidesNum > pptxMaxSlides;
+                                }
+                            }
+                            break;
+                        case MimetypeMap.MIMETYPE_PPT:
+                            if (pptMaxSlides > 0) {
+                                try (InputStream inputStream = contentReader.getContentInputStream()) {
+                                    HSLFSlideShow ppt = new HSLFSlideShow(inputStream);
+                                    int slidesCount = ppt.getDocumentSummaryInformation().getSlideCount();
+                                    logger.debug("PPT slides count is: {}", slidesCount);
+                                    result = slidesCount > pptMaxSlides;
+                                }
+                            }
+                        default:
+                            break;
+                    }
+                } catch (Exception e) {
+                    logger.error("Get input stream of file failed, can't parse message: {}", e.getMessage());
+                }
+            }
+        }
+        return result;
+    }
 }
+