+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.5] 2016-12-14
+
+### Changed
+
+* Documents that are above preview thresholds will no longer be previewed in Share
+* Thresholds values can be set accordingly in alfresco-global.properties, the example settings are:
+```
+  onlyoffice.preview.document.size.threshold=10485760
+  onlyoffice.preview.docx.threshold=8000
+  onlyoffice.preview.doc.threshold=8000
+  onlyoffice.preview.xlsx.threshold=10000
+  onlyoffice.preview.xls.threshold=10000
+  onlyoffice.preview.pptx.threshold=1000
+  onlyoffice.preview.ppt.threshold=1000
+```
+* In the example settings above, document size threshold is 10Mb, the docx and doc threshold is max paragraphs number, the xlsx and xls threshold is max total rows of sheets, the pptx and ppt thresholds is max slides number.
+* Any threshold's value will be 0 if they are missing in properties file. Therefore the corresponding threshold check won't be performed. For example, if `onlyoffice.preview.document.size.threshold` is not present in alfresco-global.properties, then there won't be any limitation in terms of document size when user tries to preview a document.
+
 ## [1.2.4] 2016-11-23
 
 ### Fixed